[multi forge] move isFile from env folder to forge-fs
