+++ src/cli/project/create/create.js
@@ -5,9 +5,9 @@ const fuzzy = require('fuzzy');
 const inquirer = require('inquirer');
 const shell = require('shelljs');
 const GraphQLClient = require('@arcblock/graphql-client');
-const { isFile, webUrl } = require('core/env');
+const { webUrl } = require('core/env');
 const debug = require('core/debug')('create');
-const { isDirectory } = require('core/forge-fs');
+const { isDirectory, isFile } = require('core/forge-fs');
 const { symbols, hr } = require('core/ui');
 inquirer.registerPrompt('autocomplete', require('inquirer-autocomplete-prompt'));
