+++ src/cli/protocol/compile/compile.js
@@ -4,9 +4,9 @@ const shell = require('shelljs');
 const chalk = require('chalk');
 const yaml = require('yaml');
 const { symbols, getSpinner } = require('core/ui');
-const { isFile, DEFAULT_MIRROR } = require('core/env');
+const { DEFAULT_MIRROR } = require('core/env');
 const debug = require('core/debug')('compile');
-const { isDirectory, requiredDirs } = require('core/forge-fs');
+const { isFile, isDirectory, requiredDirs } = require('core/forge-fs');
 const { downloadAsset } = require('../../node/install/install');
 
 // eslint-disable-next-line consistent-return
