+++ src/core/env.js
@@ -76,10 +76,6 @@ async function setupEnv(args, requirements) {
   }
 }
 
-function isFile(x) {
-  return fs.existsSync(x) && fs.statSync(x).isFile();
-}
-
 function isEmptyDirectory(x) {
   return isDirectory(x) && fs.readdirSync(x).length === 0;
 }
@@ -464,7 +460,6 @@ module.exports = {
   getPlatform,
   createRpcClient,
   isDirectory,
-  isFile,
   isEmptyDirectory,
   printLogo,
   ensureConfigComment,
